Revert "Bugfix/modularizacion app clientes"
