--- chats_IA/sprint4/chatISAAC.txt
@@ -1,1 +1,0 @@
-- Isaac Gonzales:
