--- chats_IA/sprint4/chatIVAN.txt
@@ -1,2 +1,0 @@
-- Ivan Figueredo:
-
