--- chats_IA/sprint4/chatJOSE.txt
@@ -1,1 +1,0 @@
-- Jose Zaavedra:
