--- chats_IA/sprint4/chatLEO.txt
@@ -1,1 +1,0 @@
-- Leonardo Rios:
