--- chats_IA/sprint4/chatNICO.txt
@@ -1,1 +1,0 @@
-- Nicolas Arza:
