+++ clientes/templates/clientes/medios_pago/form_medio_pago.html
@@ -82,14 +82,13 @@
 
                         <!-- Campos básicos del medio de pago -->
                         <div class="row mb-4">
-                            <!-- Tipo de Medio de Pago -->
-                            <div class="col-md-6">
+                            <div class="col-md-8">
                                 <div class="mb-3">
-                                    <label class="form-label fw-bold">
-                                        <i class="fas fa-credit-card me-2 text-primary"></i>Tipo de Medio de Pago
+                                    <label class="form-label">
+                                        <i class="fas fa-credit-card me-2"></i>Tipo de Medio de Pago
                                     </label>
                                     <div class="p-3 bg-light border rounded-3 d-flex align-items-center">
-                                        <i class="fas fa-credit-card fa-2x text-primary me-3"></i>
+                                        <i class="fas fa-credit-card fa-lg text-primary me-3"></i>
                                         <div>
                                             <div class="fw-bold text-dark">{{ medio_de_pago.nombre }}</div>
                                             <small class="text-muted">Medio de pago seleccionado</small>
@@ -98,31 +97,25 @@
                                     {{ form.medio_de_pago.as_hidden }}
                                 </div>
                             </div>
-
-                            <!-- Estado y Configuración -->
-                            <div class="col-md-6">
-                                <label class="form-label fw-bold">
-                                    <i class="fas fa-cog me-2 text-secondary"></i>Estado y Configuración
-                                </label>
-                                
-                                <div class="p-3 bg-light border rounded-3">                
-
-                                    <!-- Switch: Principal -->
+                            <div class="col-md-4">
+                                <div class="mb-3">
+                                    <label for="{{ form.es_principal.id_for_label }}" class="form-label">
+                                        <i class="fas fa-star me-2"></i>Configuración
+                                    </label>
                                     <div class="form-check form-switch">
                                         {{ form.es_principal }}
-                                        <label class="form-check-label fw-semibold" for="{{ form.es_principal.id_for_label }}">
-                                            <i class="fas fa-star me-2 text-warning"></i>
-                                            Medio de pago principal
+                                        <label class="form-check-label" for="{{ form.es_principal.id_for_label }}">
+                                            Marcar como medio principal
                                         </label>
-                                        <small class="d-block text-muted mt-1">
-                                            Solo un medio puede ser principal por cliente
-                                        </small>
-                                        {% if form.es_principal.errors %}
-                                            <div class="text-danger small mt-2">
-                                                {{ form.es_principal.errors }}
-                                            </div>
-                                        {% endif %}
                                     </div>
+                                    <div class="form-text">
+                                        El medio principal se usará por defecto en las transacciones.
+                                    </div>
+                                    {% if form.es_principal.errors %}
+                                        <div class="text-danger small mt-1">
+                                            {{ form.es_principal.errors }}
+                                        </div>
+                                    {% endif %}
                                 </div>
                             </div>
                         </div>
