+++ clientes/tests/test_limites_y_asociacion.py
@@ -1,6 +1,5 @@
-"""
-Pruebas para límites y asociaciones.
-"""
+# clientes/tests/test_limites_y_asociacion.py (VERSIÓN FINAL CON IMPRESIÓN DE ERRORES CLAROS)
+
 from django.test import TestCase
 from django.urls import reverse
 from django.contrib.auth import get_user_model
@@ -11,10 +10,10 @@ from unittest.mock import patch, MagicMock, PropertyMock
 from django.contrib.contenttypes.models import ContentType
 from django.contrib.auth.models import Permission
 
-# ✅ Imports modularizados
+# Importaciones
 from clientes.models import Cliente, AsignacionCliente, Segmento, LimiteDiario, LimiteMensual
 from clientes.services import verificar_limites
-from clientes.forms.limite import LimiteDiarioForm, LimiteMensualForm  # ✅ CORREGIR
+from clientes.forms import LimiteDiarioForm, LimiteMensualForm # Necesarias para los tests de validación
 
 User = get_user_model()
 
@@ -368,10 +367,11 @@ class TestModelosBasicos(TestCase):
         form = LimiteDiarioForm(data=data)
         
         if not form.is_valid():
+            # ⬇️ IMPRESIÓN MEJORADA
             print(f"\n🚨 Errores de Formulario (Duplicado Diario):\n{dict(form.errors)}")
         self.assertFalse(form.is_valid())
         self.assertIn('fecha', form.errors)
-        self.assertIn("Ya existe un límite configurado para", form.errors['fecha'][0])  # ✅ CORREGIDO
+        self.assertIn("Ya existe un límite configurado para esta fecha.", form.errors['fecha'][0])
 
 
     @patch('django.utils.timezone.localdate')
@@ -394,10 +394,11 @@ class TestModelosBasicos(TestCase):
         form = LimiteMensualForm(data=data)
         
         if not form.is_valid():
+            # ⬇️ IMPRESIÓN MEJORADA
             print(f"\n🚨 Errores de Formulario (Duplicado Mensual):\n{dict(form.errors)}")
         self.assertFalse(form.is_valid())
         self.assertIn('mes', form.errors)
-        self.assertIn("Ya existe un límite configurado para", form.errors['mes'][0])  # ✅ CORREGIDO
+        self.assertIn("Ya existe un límite configurado para este mes.", form.errors['mes'][0])
 
 
     def test_limite_diario_monto_negativo_falla(self):
@@ -411,12 +412,15 @@ class TestModelosBasicos(TestCase):
         form = LimiteDiarioForm(data=data)
         
         if not form.is_valid():
+            # ⬇️ IMPRESIÓN MEJORADA
             print(f"\n🚨 Errores de Formulario (Monto Negativo Diario):\n{dict(form.errors)}")
         self.assertFalse(form.is_valid())
         self.assertIn('monto', form.errors)
-        # ✅ CORREGIDO: Convertir a string y buscar el mensaje
-        error_text = str(form.errors['monto'])
-        self.assertIn("El monto debe ser mayor a cero", error_text)
+        expected_error_found = any(
+            "mayor o igual a 0" in err or "greater than or equal to 0" in err
+            for err in form.errors['monto']
+        )
+        self.assertTrue(expected_error_found, f"Mensaje de error inesperado: {form.errors['monto']}")
 
     
     def test_limite_diario_monto_vacio_falla(self):
