+++ clientes/tests/tests_clientes.py
@@ -1,13 +1,10 @@
-"""
-Pruebas unitarias para modelos y vistas de clientes.
-"""
+# clientes/tests/tests_clientes.py (CORREGIDO)
 from django.test import TestCase
-from django.contrib.auth import get_user_model  # ✅ AGREGAR
+from django.contrib.auth import get_user_model
+from clientes.models import Cliente, Segmento
+# Importaciones necesarias
 from django.db.utils import IntegrityError 
 from django.db import transaction 
-
-# ✅ Imports modularizados (ya están correctos)
-from clientes.models import Cliente, Segmento, AsignacionCliente
 
 CustomUser = get_user_model()
